refactor: Moved item functionality into item service from Inventory Service
