+++ src/main/frontend/ng-food-inventory/src/components/inventory/edit-item-form/edit-item-form.component.ts
@@ -1,7 +1,7 @@
 import {Component, ElementRef, Input, OnInit, ViewChild, inject} from '@angular/core';
 import { FormBuilder, FormControl, FormGroup, Validators, ReactiveFormsModule } from "@angular/forms";
 import {Item} from "../../../models/item.model";
-import {InventoryService, LocationService} from "../../../services";
+import {InventoryService, ItemService, LocationService} from "../../../services";
 import { AddLocationModalComponent } from '../add-location-modal/add-location-modal.component';
 import { NgIf, NgFor } from '@angular/common';
 
@@ -21,7 +21,8 @@ export class EditItemFormComponent implements OnInit {
 
   itemForm: FormGroup|undefined;
   showModal = false;
-
+  
+  itemService = inject(ItemService);
   inventoryService = inject(InventoryService);
   locationService = inject(LocationService);
   fb = inject(FormBuilder);
@@ -54,7 +55,7 @@ export class EditItemFormComponent implements OnInit {
     this.item.quantity = value.quantity;
     this.item.location = this.locationService.getLocation(value.location);
 
-    this.inventoryService.saveItem(this.item).subscribe({
+    this.itemService.saveItem(this.item).subscribe({
       error: console.error
     });
   }
